typos and suggestions
Typos.
Add consistency to variable naming style in examples
Remove use of simple profile from examples which occur before profiles have been explained and where using it does not add to the utility of the example.
Use colon as version string separator in profile examples so that they match the guidance.
Deleted a reference which implied TOSCA can only be used for cloud applications.
remove references to TOSCA profile from examples where it adds no value
mark sections as normative or not
Improve use of language for clarity
Add references
extend an example